refactor(components): adjust some file names and imports
related to #eu colocaria o task-id se tivesse integracao algum plataforma (jira, az devops, etc).
